requirements.txt
asgiref==3.8.1
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==5.1.4
django-cors-headers==4.6.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.2.1
openapi-codec==1.3.2
packaging==24.2
pillow==10.4.0
proglog==0.1.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
uritemplate==4.1.1
urllib3==2.3.0
